+++ packages/nimble-components/.storybook/preview.js
@@ -15,7 +15,21 @@ export const parameters = {
     options: {
         storySort: {
             method: 'alphabetical',
-            order: ['Getting Started']
+            // Items within arrays show the sort order for children of the category above
+            // https://storybook.js.org/docs/react/writing-stories/naming-components-and-hierarchy#sorting-stories
+            order: [
+                'Getting Started',
+                'Components',
+                ['Status Table'],
+                'Incubating',
+                ['Docs'],
+                'Tokens',
+                ['Docs'],
+                'Tests',
+                ['Docs'],
+                'Internal',
+                ['Docs']
+            ]
         }
     },
     controls: {