+++ packages/nimble-components/src/all-components.ts
@@ -28,6 +28,7 @@ import './menu-item';
 import './number-field';
 import './radio';
 import './radio-group';
+import './rich-text-viewer';
 import './select';
 import './spinner';
 import './switch';