+++ packages/nimble-components/src/table-column/base/template.ts
@@ -1,10 +1,16 @@
-import { html } from '@microsoft/fast-element';
+import { html, ref } from '@microsoft/fast-element';
 import type { TableColumn } from '.';
+import { overflow } from '../../utilities/directive/overflow';
 
+// prettier-ignore
 export const template = html<TableColumn>`
     <template slot="${x => x.columnInternals.uniqueId}">
-        <span class="header-content">
-            <slot></slot>
+        <span
+            ${overflow('hasOverflow')}
+            class="header-content"
+            title=${x => (x.hasOverflow && x.headerTextContent ? x.headerTextContent : null)}
+        >
+            <slot ${ref('contentSlot')}></slot>
         </span>
     </template>
 `;