+++ packages/nimble-components/src/table-column/text-base/cell-view/index.ts
@@ -11,31 +11,25 @@ export abstract class TableColumnTextCellViewBase<
 > extends TableCellView<TCellRecord, TColumnConfig> {
     /** @internal */
     @observable
-    public isValidContentAndHasOverflow = false;
-
-    /** @internal */
-    public textSpan!: HTMLElement;
+    public hasOverflow = false;
 
     /**
-     * Returns the text to render in the cell when it contains a valid value (i.e. when shouldUsePlaceholder() is false).
-     * If the implementation has branching code paths then it must be marked with @volatile.
-     * https://www.fast.design/docs/fast-element/observables-and-state/#observable-features
+     * Text to render in the cell when it contains a valid value (i.e. when shouldUsePlaceholder is false).
      */
-    public abstract get text(): string;
+    @observable
+    public text = '';
 
     /**
-     * Returns the text to render in the cell when it contains an invalid value (i.e. when shouldUsePlaceholder() is true).
-     * If the implementation has branching code paths then it must be marked with @volatile.
-     * https://www.fast.design/docs/fast-element/observables-and-state/#observable-features
+     * Text to render in the cell when it contains an invalid value (i.e. when shouldUsePlaceholder is true).
      */
-    public abstract get placeholder(): string;
+    @observable
+    public placeholder = '';
 
     /**
      * Returns whether to display the placeholder value or the text value
-     * If the implementation has branching code paths then it must be marked with @volatile.
-     * https://www.fast.design/docs/fast-element/observables-and-state/#observable-features
-     * */
-    public abstract get shouldUsePlaceholder(): boolean;
+     */
+    @observable
+    public shouldUsePlaceholder = true;
 
     @volatile
     public get content(): string {